Update consumers.py
one thing im not sure about is the user creation, i've not given it a password and in the documentation it says this means the user will not be saved. i dont know if that means it can be used as a one time only thing or it just wont work. otherwise each bot needs a password which we could just generate randomly